more comments added, reinforced React components model, refactored CSS, added cleanup functions to useEffects
