+++ .gitignore
@@ -5,5 +5,9 @@
 /build
 /dist
 
+# misc
+.DS_Store
+.vscode
+
