--- .node-version
@@ -1,1 +1,0 @@
-16.16.0
