--- src/components/Navigation.tsx
@@ -1,6 +1,0 @@
-// forest (will also editing the app page to add react router)
-import '../styles/navigation.css';
-
-export default {
-  
-}
