+++ src/components/smallComps/NoPage.tsx
@@ -2,14 +2,11 @@ import React from "react";
 
 const NoPage = () => {
 
-return (
-
-    <div>
-        <p>No Page Found, try again</p>
-    </div>
-
-)
-
+    return (
+        <div>
+            <p>No Page Found, try again</p>
+        </div>
+    )
 
 }
