+++ src/components/statelessComps/Display.tsx
@@ -1,0 +1,15 @@
+// importing dependencies 
+import React from 'react';
+
+// component for displaying modals based on active pod 
+const Display = (props: any) => {
+
+  return(
+    <div className='display'>
+      <h5>fake graphs lol</h5>
+      <img src="https://miro.medium.com/max/1400/1*QwGqOMObJd7oFHCfe1AvxA.png" alt="" />
+    </div>
+  )
+}
+
+export default Display;
