+++ src/styles/dashboard.css
@@ -1,37 +1,20 @@
 .dashboard {
   width: 100%;
   height: 200%;
-  background-size: cover;
 }
 
 .top-bar {
-  /* border: 2px solid black;
-  background-color: #EBF0E5; */
   position: fixed;
   top: 0;
   display: flex;
-  flex-direction: row;
-  background-size: 200% auto;
-  align-content: center;
   align-items: center;
   justify-content: space-around;
-  border: none;
-  box-shadow: 0 4px 6px #E2E0E1, 0 1px 3px #EC6536;
-  background-image: linear-gradient(to right, #ec6336a4 0%, #f7bd58df 50%, #ec6336a4 100%);
   transition: 0.5s;
-  display: flex;
-  justify-content: center;
   width: 100%;
   height: 75px;
+  border: 1.2px solid #ec633677;
+  box-shadow: 0 2px 3px #E2E0E1, 0 1px 3px #ec6336b5;
 }
-
-.top-bar:hover {
-  background-position: right center;
-}
-
-/* .logo {
-  margin-right: 200px;
-} */
 
 .header-logo {
   position: fixed;
@@ -79,68 +62,44 @@
 
 .pods-header {
   height: 35px;
-  width: 96%;
-  /* position: fixed; */
-  /* position: absolute; */
-
+  width: 100%;
   display: flex;
-  /* justify-content: center; */
-  
-  background-size: 200% auto;
   border: none;
-  border-radius: 1rem;
-  box-shadow: 0 4px 6px #f7bd58df, 0 1px 3px #EC6536;
-  background-image: linear-gradient(to right, #ec6336a4 0%, #f7bd58df 50%, #ec6336a4 100%);
+  border-radius: 5px;
+  background-color: rgb(222, 145, 41);
   transition: 0.5s;
-  /* margin-top: 5px;
-  margin-bottom: 5px; */
   margin-top: 30px;
   margin-bottom: -15px;
-  /* margin-bottom: 50px; */
 }
 
 .pods-header h5 {
   padding-left: 10px;
   margin: 30px;
   margin-top: 10px;
+  color: white;
 }
 
 .all-pod-info {
   border: 1.2px solid #ec633677;
-  border-radius: 1rem;
-  box-shadow: 0 4px 6px #E2E0E1, 0 1px 3px #ec6336b5;
+  border-radius: 5px;
   flex: 0.3;
-
   height: 290px;
   width: 650px;
-
   margin-top: 20px;
-  overflow-y: scroll;
-  
+  overflow-y: auto
 }
 
 .pod-info {
   height: 50px;
   width: 97%;
-
   display: flex;
   justify-content: center;
-  
   background-color: #E2E0E1;
   border: 1.2px solid #ec633677;
-  border-radius: 1rem;
-  box-shadow: 0 4px 6px #E2E0E1, 0 1px 3px #ec6336b5;
-  /* margin-top: 5px;
-  margin-bottom: 5px; */
+  border-radius: 5px;
   margin: 5px;
   padding-bottom: 10px;
   padding-top: -10px;
-}
-
-
-
-.pods-header:hover {
-  background-position: right center;
 }
 
 .pod-info h5 {
@@ -172,30 +131,5 @@
   .pods-header {
     width: 190px;
   }
-}
-
-
-.url-input {
-  display: flex;
-  justify-content: center;
-  text-align: center;
-  height: 50%;
-  margin-left: 20px;
-  margin-right: 20px;
-  align-self: center;
-}
-
-.url-input h5 {
-  margin-top: 5px;
-  color: white;
-  font-weight: 800;
-    font-size: 1em;
-    text-transform: uppercase;
-}
-
-.login-buttons{
-  display: flex;
-  align-items: center;
-  margin-left: 100px;
 }
